+++ CHANGELOG.md
@@ -119,6 +119,8 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Add Image Text 25/75 and Full Width Text into SublandingPage
 - Add related_posts_function to the global context in Jinja2 for prototyping of related posts
 - Added the featured content module molecule and included it in the landing-page prototype
+- Add ImageText2575Group organism
+- Add ImageText2575Group to Sublanding and Landing pages 
 
 ### Changed
 - Updated the primary nav to move focus as user enters and leaves nav levels
@@ -200,6 +202,8 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - Temporarily comment out related posts section of single blog post browser test until BlogPage's are in Wagtail
 - Add `show_heading` checkbox to Related Posts organism to toggle the heading
   and icon.
+- Merge Streamfields in LandingPage
+- Landing and Sublanding content blocks render each atomic structure with `div class="block">`
 
 ### Removed
 - Removed unused exportsOverride section,